Minor fixes.

+++ DuplicateFinder/src/main.cpp
@@ -3,9 +3,9 @@
 int main( int argc, char** argv )
 {
     std::vector<std::string> arguments;
-    for( size_t i = 0; i < argc; ++i )
+    for( int i = 0; i < argc; ++i )
     {
-        arguments.push_back( std::string( argv[i] ) );
+        arguments.push_back( std::string( argv[(size_t)i] ) );
     }
 
     std::string searchPath = ".";
